+++ packages/terra-compact-interactive-list/src/proptypes/unitShape.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+const unitShape = PropTypes.shape({
+  value: PropTypes.number,
+  unitTupe: PropTypes.string,
+});
+
+export default unitShape;