+++ packages/terra-data-grid/src/DataGrid.jsx
@@ -195,6 +195,8 @@ const DataGrid = forwardRef((props, ref) => {
   const tableContainerRef = useRef();
   const handleFocus = useRef(true);
 
+  const focusedCellRef = useRef({ rowId: '', columnId: '' });
+
   const [checkResizable, setCheckResizable] = useState(false);
 
   // if columns are not visible then set the first selectable row index to 1
@@ -215,14 +217,19 @@ const DataGrid = forwardRef((props, ref) => {
   // -------------------------------------
   // functions
 
-  const isRowSelectionCell = (columnIndex) => (
+  const isRowSelectionCell = useCallback((columnIndex) => (
     hasSelectableRows && columnIndex < displayedColumns.length && displayedColumns[columnIndex].id === WorklistDataGridUtils.ROW_SELECTION_COLUMN.id
-  );
+  ), [displayedColumns, hasSelectableRows]);
 
-  const setFocusedRowCol = (newRowIndex, newColIndex, makeActiveElement) => {
+  const setFocusedRowCol = useCallback((newRowIndex, newColIndex, makeActiveElement) => {
     setCellAriaLiveMessage(null);
     setFocusedRow(newRowIndex);
     setFocusedCol(newColIndex);
+
+    focusedCellRef.current = {
+      rowId: grid.current.rows[newRowIndex].getAttribute('data-row-id'),
+      columnId: displayedColumns[newColIndex].id,
+    };
 
     if (makeActiveElement) {
       // Set focus on input field (checkbox) of row selection cells.
@@ -238,7 +245,7 @@ const DataGrid = forwardRef((props, ref) => {
 
       focusedCell?.focus();
     }
-  };
+  }, [isRowSelectionCell, displayedColumns]);
 
   // The focus is handled by the DataGrid. However, there are times
   // when the other components may want to change the currently focus
@@ -295,28 +302,27 @@ const DataGrid = forwardRef((props, ref) => {
 
   const handleColumnSelect = useCallback((columnId) => {
     const columnIndex = displayedColumns.findIndex(column => column.id === columnId);
-    setFocusedCol(columnIndex);
+    setFocusedRowCol(0, columnIndex);
 
     if (onColumnSelect) {
       onColumnSelect(columnId);
     }
-  }, [onColumnSelect, displayedColumns]);
+  }, [onColumnSelect, displayedColumns, setFocusedRowCol]);
 
   const handleRowSelectionHeaderSelect = useCallback(() => {
-    setFocusedCol(0);
-    setFocusedRow(0);
+    setFocusedRowCol(0, 0);
     if (onRowSelectionHeaderSelect) {
       onRowSelectionHeaderSelect();
     }
-  }, [onRowSelectionHeaderSelect]);
+  }, [onRowSelectionHeaderSelect, setFocusedRowCol]);
 
   const handleCellSelection = useCallback((selectionDetails) => {
-    setFocusedRow(selectionDetails.rowIndex);
-    setFocusedCol(selectionDetails.columnIndex);
+    const { columnIndex, rowIndex } = selectionDetails;
+    setFocusedRowCol(rowIndex, columnIndex);
     if (onCellSelect) {
       onCellSelect(selectionDetails);
     }
-  }, [onCellSelect]);
+  }, [onCellSelect, setFocusedRowCol]);
 
   // -------------------------------------
   // event handlers
@@ -478,6 +484,7 @@ const DataGrid = forwardRef((props, ref) => {
       event.preventDefault(); // prevent the page from moving with the arrow keys.
       return;
     }
+
     handleMoveCellFocus(cellCoordinates, { row: nextRow, col: nextCol });
     event.preventDefault(); // prevent the page from moving with the arrow keys.
   };
@@ -495,7 +502,26 @@ const DataGrid = forwardRef((props, ref) => {
     if (!event.currentTarget.contains(event.relatedTarget)) {
       // Not triggered when swapping focus between children
       if (handleFocus.current) {
-        setFocusedRowCol(focusedRow, focusedCol, true);
+        let newRowIndex = focusedRow;
+        let newColumnIndex = focusedCol;
+
+        // Check for last focused row ID. If found set the index. Otherwise set it to the last focused row or last index.
+        if (focusedCellRef.current.rowId) {
+          newRowIndex = [...grid.current.rows].findIndex(row => row.getAttribute('data-row-id') === focusedCellRef.current.rowId);
+          newRowIndex = newRowIndex === -1
+            ? Math.min(focusedRow, grid.current.rows.length - 1)
+            : newRowIndex;
+        }
+
+        // Check for last focused column ID. If found set the index. Otherwise set it to the last focused column or last index.
+        if (focusedCellRef.current.columnId) {
+          newColumnIndex = displayedColumns.findIndex(column => column.id === focusedCellRef.current.columnId);
+          newColumnIndex = newColumnIndex === -1
+            ? Math.min(focusedCol, displayedColumns.length - 1)
+            : newColumnIndex;
+        }
+
+        setFocusedRowCol(newRowIndex, newColumnIndex, true);
         setGridHasFocus(true);
       }
     }
@@ -555,7 +581,7 @@ const DataGrid = forwardRef((props, ref) => {
   );
 });
 
-DataGrid.propTypes = propTypes;
 DataGrid.defaultProps = defaultProps;
+DataGrid.propTypes = propTypes;
 
 export default DataGrid;